# src/configs/swagger.yaml
openapi: 3.0.3
info:
  title: 'Tandem Proyect 2026 - API Backend'
  version: 1.0.0
  description: |
    # Tándem 2026 — API Backend

    API del proyecto Tándem, una plataforma de apoyo para personas con TEA (Trastorno del Espectro Autista) que integra
    gestión de usuarios, actividades, gamificación, comunicación en tiempo real y seguimiento de ubicación.

    ## Stack principal
    - **Runtime:** Node.js + Express 5
    - **Base de datos:** PostgreSQL con SQL crudo (sin ORM)
    - **Tiempo real:** Socket.io 4 (chat y notificaciones)
    - **Caché / Colas:** Redis + ioredis + BullMQ (opcional)
    - **Autenticación:** JWT custom HMAC-SHA256 + Refresh Tokens + CSRF double-submit cookie
    - **APIs externas:** ARASAAC (pictogramas), DiceBear (avatares), Supabase Storage

    ## Arquitectura
    ```
    Controller → Service → Repository → DB (PostgreSQL)
                    ↓
          CacheService (Redis) / Socket.io / APIs externas
    ```

    Middleware chain: `helmet → cors → cookieParser → rate-limit → authMiddleware (JWT) → csrfMiddleware → Controller`

    ## Documentación visual de arquitectura
    Visitá [/docs](/docs) para ver diagramas interactivos de la base de datos, flujo de autenticación,
    chat realtime, Redis, APIs externas y modelo de permisos.
servers:
  -
    url: /api
    description: 'Servidor API'
  -
    url: http://localhost:3000/api
    description: 'Servidor de Desarrollo Local'
...
security:
  -
    CookieAuth: []
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: tandem_access_token
  schemas:
...
    Perteneciente:
      type: object
      properties:
        id:
          type: integer
          example: 2
        id_usuario:
          type: integer
          example: 7
        id_nivel_apoyo:
          type: integer
          example: 10
        id_autonomia_operativa:
          type: integer
          example: 3
        puede_autogestionarse:
          type: boolean
          example: true
        observacion_general:
          type: string
          example: 'Descripción detallada para Perteneciente'
...
paths:
...
  /pertenecientes:
    get:
      summary: 'Listar pertenecientes'
      tags:
        - 'Módulo Perteneciente'
      responses:
        '200':
          description: 'Operación exitosa'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Perteneciente'
    post:
      summary: 'Crear Perteneciente'
      tags:
        - 'Módulo Perteneciente'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Perteneciente'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: 'Creado exitosamente'
  '/pertenecientes/{id}':
    get:
      summary: 'Obtener Perteneciente por ID'
      tags:
        - 'Módulo Perteneciente'
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Operación exitosa'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perteneciente'
        '404':
          description: 'No encontrado'
    put:
      summary: 'Actualizar Perteneciente por ID'
      tags:
        - 'Módulo Perteneciente'
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Perteneciente'
      responses:
        '200':
          description: 'Actualizado correctamente'
        '404':
          description: 'No encontrado'
    delete:
      summary: 'Eliminar Perteneciente por ID'
      tags:
        - 'Módulo Perteneciente'
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Eliminado correctamente'
        '404':
          description: 'No encontrado'
...